+++ CHANGELOG.md
@@ -11,3 +11,4 @@ FEATURES:
 * **New Data Source:** `proxmox_virtual_environment_version`
 * **New Resource:** `proxmox_virtual_environment_group`
 * **New Resource:** `proxmox_virtual_environment_role`
+* **New Resource:** `proxmox_virtual_environment_user`